+++ example/ini_file_example.dart
@@ -1,9 +1,15 @@
 import 'package:ini_file/ini_file.dart';
 
 void main() async {
-  IniFile ini = IniFile();
+  // create instance and eliminate spaces from results
+  // key = abc will return:
+  // trimResults = true ?  'abc'
+  // trimResults = false ? ' abc'
+  IniFile ini = IniFile(trimResults: true);
 
   await ini.readFile('./example/YSA.ini');
+
+  print(ini.toMap());
 
   print(ini.getItems('SETTINGS'));
   print(ini.getJsonItems('SETTINGS'));
@@ -23,4 +29,9 @@ void main() async {
   ini.removeSection('SETTINGS');
   ini.setItem('NEWSECTION', 'newitem', 'new value');
   print(ini.getItems('NEWSECTION'));
+
+  ini = IniFile.emptyFile('./example/new.ini', overrideFile: true);
+  ini.setItem('section', 'key', 'value');
+  ini.writeFile();
+  print(ini.getItem('section', 'key'));
 }